backend/config-service/requirements.txt
# FastAPI和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis缓存
redis[hiredis]==5.0.1

# 认证和安全
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# 配置文件格式支持
pyyaml==6.0.1
toml==0.10.2
configparser==6.0.0

# JSON Schema验证
jsonschema==4.20.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 工具库
python-multipart==0.0.6
python-dateutil==2.8.2
click==8.1.7
rich==13.7.0

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 类型提示
types-PyYAML==6.0.12.12
types-redis==4.6.0.11
types-passlib==1.7.7.13